backend/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
pydantic==1.7.3
ujson==4.0.2
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
requests==2.25.1

# to read pyproject.toml configuration
toml==0.10.2

# better terminal
ipython==7.19.0

# for unit tests
pytest==6.2.2
factory-boy==3.2.0
pytest-mock==3.5.1

# for formatting and linters
isort==5.7.0
black==20.8b1
flake8==3.8.4
flake8-bugbear==20.11.1

# mypy related
mypy==0.800
